+++ client/README.md
@@ -4,5 +4,4 @@ App.js contains my form code that triggers fetch request to the Lambda URL
 index.js contains Lambda logic
 \*\*\* ses allows us to write sendMail func.
 
-1. emailParams object
-2. tryCatch block
+We're packaging up data from a form and submitting it in the 'sendMail' function within a params object
