--- client/src/Loading.jsx
@@ -1,11 +1,0 @@
-const Loading = () => {
-  return (
-    <div className="flex justify-center items-center">
-      <div className="mt-30">
-        <p className="text-5xl animate-pulse">LOADING</p>
-      </div>
-    </div>
-  )
-}
-
-export default Loading
